wasm: cast Poisson() return from ULong64_t to int for JS compatibility
Emscripten embind maps ULong64_t (unsigned long long) to JavaScript BigInt,
which cannot be used with Math.trunc() or compared with ===. Cast the result
to int in the binding lambda since Poisson counts for typical means easily
fit in a 32-bit integer.

+++ core/wasm/src/RootWasmBindings.cxx
@@ -180,7 +180,7 @@ EMSCRIPTEN_BINDINGS(TRandom3_bindings)
          return r.Uniform(lo, hi);
       }))
       .function("Poisson", optional_override([](TRandom3 &r, double mean) {
-         return r.Poisson(mean);
+         return static_cast<int>(r.Poisson(mean));  // ULong64_t → int for JS
       }))
       .function("SetSeed", optional_override([](TRandom3 &r, unsigned int seed) {
          r.SetSeed(seed);
